backend/requirements.txt
Flask==3.0.3
Flask-SQLAlchemy==3.1.1
Flask-Migrate==4.0.7
Flask-JWT-Extended==4.6.0
Flask-CORS==4.0.1
Flask-Limiter==3.8.0
bcrypt==4.2.0
python-dotenv==1.0.1
marshmallow==3.22.0
marshmallow-sqlalchemy==1.1.0
